clean and splitted
