--- src/graphics/mod.rs
@@ -1,2 +1,0 @@
-pub mod visual;
-
