+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod board;
 pub mod args;
 pub mod algo;
-pub mod graphics;
+pub mod visual;
