+++ src/visual/mod.rs
@@ -1,0 +1,5 @@
+pub mod render;
+pub mod text;
+pub mod visu;
+
+
