+++ src/visual/visu.rs
@@ -1,0 +1,104 @@
+extern crate piston_window;
+extern crate opengl_graphics;
+extern crate find_folder;
+
+use crate::board::utils::{fdtos};
+use crate::visual::text::{DrawText, TextAlignment, TextVerticalAlignment};
+
+use opengl_graphics::{GlGraphics, GlyphCache};
+use piston_window::*;
+
+pub struct Visu {
+    pub gl: GlGraphics,
+    pub board: Vec<i32>,
+    pub size: i32,
+    pub time: String,
+    pub margin_top: f64,
+    pub margin_x: f64,
+    pub number_scale: f64,
+    pub index: i32,
+    pub total_moves: i32,
+}
+
+impl Visu {
+    pub fn render(&mut self, args: &RenderArgs) {
+        const RED: [f32; 4] = [1.0, 0.0, 0.0, 1.0];
+        const LIGHT_GREY: [f32; 4] = [218.0 / 255.0, 223.0 / 255.0 , 225.0 / 255.0, 1.0];
+        const DARK_GREY: [f32; 4] = [46.0 / 255.0, 49.0 / 255.0, 49.0 / 255.0, 1.0];
+
+        let size = self.size;
+        let margin_top = self.margin_top;
+        let margin_x = self.margin_x;
+        let number_scale = self.number_scale;
+
+        let (win_w, win_h) = (args.window_size[0], args.window_size[1]);
+
+        let grid = grid::Grid {
+            cols: size as u32,
+            rows: size as u32,
+            units: win_w / size as f64 - (margin_x * 2.0 / size as f64),
+        };
+
+        let line = Line::new(RED, 1.5);
+        
+        let board = self.board.clone();
+        let time = self.time.clone();
+        let total_moves = self.total_moves;
+        let index = self.index;
+
+        let assets = find_folder::Search::ParentsThenKids(3, 3)
+        .for_folder("assets").unwrap();
+        let ref font = assets.join("font.ttf");
+        let mut glyph_cache = GlyphCache::new(font, (), TextureSettings::new()).unwrap();
+
+        self.gl.draw(args.viewport(), |c, gl| {
+            clear(DARK_GREY, gl);
+                   
+            grid.draw(&line, &c.draw_state, c.transform
+                .trans(margin_x, margin_top), gl);
+
+            for y in 0..size as u32 {
+                for x in 0..size as u32 {
+                    let pos = grid.cell_position((x, y));
+                    let nb = board[fdtos(x as i32, y as i32, size) as usize];
+  
+                    if nb != size * size {
+                        let string: String = nb.to_string();
+                        let r = [pos[0] + margin_x, pos[1] + margin_top, pos[0] + margin_x + grid.units, pos[1] + margin_top + grid.units];
+                        gl.draw_text(&string, r, LIGHT_GREY, ((64.0 * (number_scale / size as f64)) as u32) as u32, TextAlignment::Center, TextVerticalAlignment::Center, &mut glyph_cache, &c);
+                    }
+                }
+            }        
+            
+            let mut r = [margin_x, 0.0, win_w - margin_x, margin_top];
+            gl.draw_text("NPUZZLE", r, LIGHT_GREY, 64, TextAlignment::Center, TextVerticalAlignment::Center, &mut glyph_cache, &c);
+            
+            let move_str = format!("{} : {}/{}", "Move", index.to_string(), total_moves.to_string());
+            r = [0.0, win_w + margin_top, win_w, win_w + margin_top + 1.0 * 35.0];
+            gl.draw_text(&move_str, r, LIGHT_GREY, 32, TextAlignment::Center, TextVerticalAlignment::Center, &mut glyph_cache, &c);
+
+            let duration_str = format!("{} : {}s", "Duration", time);
+            r = [0.0, win_w + margin_top + 1.0 * 35.0, win_w, win_w + margin_top + 2.0 * 35.0];
+            gl.draw_text(&duration_str, r, LIGHT_GREY, 32, TextAlignment::Center, TextVerticalAlignment::Center, &mut glyph_cache, &c);
+
+            r = [0.0, win_w + margin_top + 2.0 * 35.0, win_w, win_w + margin_top + 3.0 * 35.0];
+            gl.draw_text("Heuristic :", r, LIGHT_GREY, 32, TextAlignment::Center, TextVerticalAlignment::Center, &mut glyph_cache, &c);
+            
+            r = [0.0, win_w + margin_top + 3.0 * 35.0, win_w, win_w + margin_top + 4.0 * 35.0];
+            gl.draw_text("Total number of states :", r, LIGHT_GREY, 32, TextAlignment::Center, TextVerticalAlignment::Center, &mut glyph_cache, &c);
+            
+            r = [0.0, win_w + margin_top + 4.0 * 35.0, win_w, win_w + margin_top + 5.0 * 35.0];
+            gl.draw_text("Maximum number of states :", r, LIGHT_GREY, 32, TextAlignment::Center, TextVerticalAlignment::Center, &mut glyph_cache, &c);
+            
+            let moves_str = format!("{} : {}", "Number of moves", total_moves.to_string());
+            r = [0.0, win_w + margin_top + 5.0 * 35.0, win_w, win_w + margin_top + 6.0 * 35.0];
+            gl.draw_text(&moves_str, r, LIGHT_GREY, 32, TextAlignment::Center, TextVerticalAlignment::Center, &mut glyph_cache, &c);
+
+        });
+    }
+
+    pub fn update_board(&mut self, _args: &Button, board: Vec<i32>, index: usize) {
+        self.board = board;
+        self.index = index as i32;
+    }
+}
